Add test case for FileEncryption::create with filename
In order to better test FileEncryption::create, this commit adds another
test that explicitly tests it with an explicit file name for the
encrypted file. This should cover both cases where we are given either a
directory to drop off the files or an explicit filename for an encrypted
file.