Remove AWS secrets feature
In order to focus primarily on encryption/decryption of secrets, this
commit removes the AWS secrets feature to retrieve data from AWS secrets
web service. Originally the idea was to store an encryption key on AWS
and retrieve it when trying to encrypt/decrypt files. However, by
focusing on just the encryption/decryption part, we don't have to care
where the key comes from or supporting multiple cloud providers secrets
storage/retrieval platforms.