Add additional documentation for library
In order to ensure a better experience when viewing this crate on
docs.rs, this commit adds documentation to the lib.rs file. This is the
first page a user would land on when navigating to the docs. As such, it
should contain more information for the library in general and cli
usage.