Replace rust-crypto crate
In order to ensure we are using up to date and maintained dependencies,
this commit removes rust-crypto as it's no longer maintained and
replaces it with aes_gcm. Additionally it updates a few tests for
clarity of what we are testing.

Finally, it simplifies the public api of FileEncryption. It removes the
initialization vector and additional authenticated data from encrypt.
This is to ensure compatibility with existing encrypted secrets already
generated elsewhere. The initialization vector passed in was unused and
it generated it's own internally.